Fix error installing with "pip" from folder
With "pip install ./jutge-server-toolkit" an error appeared that mentioned that there is no folder "jutge/servertoolkit", corresponding to the package name. This folder is present in "jutge-toolkit" since that package has python code (whereas "jutge-server-toolkit" only has bash scripts).

I added the folder with an empty __init__.py so that the folder is now found.